[docker|game-engine] Dockerfile enhancement.
Now the Dockerfile is multi-stage.
First stage: builder. This is done only this Dockerfile was never built on the host.
Builds the rust code and produces a binary.

Second stage: runner. Copies the binary from the builder and runs it in a lighweight
container.

This configuration also aims at avoiding rights conflicts on files.

The first time (after a git clone), one has to build the game-engine locally before
running the game-engine with docker, or to comment the mount and command instructions
in the docker-compose file.